This is the final 0.2 commit.  Added allow/disallow.
Added allow and disallow predefined guards.  These guards allow
specifying a list of words that are (respectively) allowed or
disallowed.  This is useful for a set of valid values (such as scheme
for a url, like only allow "http", "https", "ftp", "ftps", etc, or
disallow certain invalid keywords that will cause errors.