Fixes #23.
This fix is a bit hacky, but it works for now.  I'd like to clean this
code up a bit... it's kind of awkward to pass the fact that the
default is prevented back to callers via a special parameter only
known internally... but I didn't really want to change any APIs in
this commit, and it was the simplest solution I could think of off the
top of my head, given those constraints.