$.fn.guard now includes selected elements, not just children.
I am going to release this as 1.5.1.  You could argue that this violates
semantic versioning, however I see this as confusing behavior that continually
bites us, so I feel it is a bug more than a new feature.  I feel it is also
unlikely to cause bugs, as including these fields seems exceedingly unlikely
to cause a bug in common usage.